unified the two Shrinking strategies
