+++ include/shark/Algorithms/QP/BoxConstrainedProblems.h
@@ -34,6 +34,7 @@
  
 #include <shark/Algorithms/QP/QpSolver.h>
 #include <shark/Algorithms/QP/Impl/AnalyticProblems.h>
+#include <shark/Algorithms/QP/BoxBasedShrinkingStrategy.h>
 
 namespace shark {
 
@@ -314,9 +315,9 @@ public:
 		}
 		setInitialSolution(alpha, gradient);
 	}
-
+	
 	///\brief Remove the i-th example from the problem.
-	virtual void deactivateVariable(std::size_t i){
+	void deactivateVariable(std::size_t i){
 		SIZE_CHECK(i < dimensions());
 		double alphai = alpha(i);
 		m_problem.alpha(i) = 0;
@@ -383,226 +384,10 @@ protected:
 		if(m_problem.alpha(i) == boxMin(i))
 			m_alphaStatus[i] |= AlphaLowerBound;
 	}
-};
-
-
-
-template<class Problem>
-struct BoxConstrainedShrinkingProblem : public BoxConstrainedProblem<Problem>{
-private:
-	typedef BoxConstrainedProblem<Problem> base_type;
-public:
-	typedef typename base_type::QpFloatType QpFloatType;
-	typedef typename base_type::MatrixType MatrixType;
-	typedef typename base_type::PreferedSelectionStrategy PreferedSelectionStrategy;
-
-	BoxConstrainedShrinkingProblem(Problem& problem, bool shrink=true)
-	: base_type(problem)
-	, m_isUnshrinked(false)
-	, m_shrink(shrink)
-	, m_gradientEdge(problem.linear)
-	{ }
-
-	using base_type::alpha;
-	using base_type::gradient;
-	using base_type::linear;
-	using base_type::active;
-	using base_type::dimensions;
-	using base_type::quadratic;
-	using base_type::isLowerBound;
-	using base_type::isUpperBound;
-	using base_type::boxMin;
-	using base_type::boxMax;
-	using base_type::setInitialSolution;
-
-	virtual void updateSMO(std::size_t i, std::size_t j){
-		double aiOld = alpha(i);
-		double ajOld = alpha(j);
-		//call base class to do the step
-		base_type::updateSMO(i,j);
-		double ai = alpha(i);
-		double aj = alpha(j);
-		
-		// update the gradient edge data structure to keep up with changes
-		updateGradientEdge(i,aiOld,ai);
-		updateGradientEdge(j,ajOld,aj);
-	}
-
-	bool shrink(double epsilon) {
-		if (! m_shrink) return false;
-
-		double largestUp;
-		double smallestDown;
-		getMaxKKTViolations(largestUp,smallestDown,active());
-
-		// check whether unshrinking is necessary at this accuracy level
-		if (!m_isUnshrinked  && (largestUp - smallestDown < 10.0 * epsilon))
-		{
-			unshrink();
-			//recalculate maximum KKT violation for immediate re-shrinking
-			getMaxKKTViolations(largestUp, smallestDown, dimensions());
-		}
-		//shrink
+	
+	bool testShrinkVariable(std::size_t a, double largestUp, double smallestDown)const{
 		smallestDown = std::min(smallestDown, 0.0);
 		largestUp = std::max(largestUp, 0.0);
-		for (std::size_t a = this->active(); a > 0; --a) {
-			if (testShrinkVariable(a-1, largestUp, smallestDown)) {
-				flipCoordinates(a-1,active()-1);
-				--this->m_active;
-			}
-		}
-		return true;
-	}
-
-	///\brief Unshrink the problem
-	void unshrink(){
-		if (active() == dimensions()) return;
-		m_isUnshrinked = true;
-		
-		// Recompute the gradient of the whole problem.
-		// We assume here that all shrinked variables are on the border of the problem.
-		// The gradient of the active components is already correct and
-		// we store the gradient of the subset of variables which are on the
-		// borders of the box for the whole set.
-		// Thus we only have to recompute the part of the gradient which is
-		// based on variables in the active set which are not on the border.
-		for (std::size_t a = active(); a < dimensions(); a++)
-			this->m_gradient(a) = m_gradientEdge(a);
-
-		for (std::size_t i = 0; i < active(); i++)
-		{
-			//check whether alpha value is already stored in gradientEdge
-			if (isUpperBound(i) || isLowerBound(i)) continue;
-			
-			QpFloatType* q = quadratic().row(i, 0, dimensions());
-			for (std::size_t a = active(); a < dimensions(); a++) 
-				this->m_gradient(a) -= alpha(i) * q[a];
-		}
-
-		this->m_active = dimensions();
-	}
-
-	void setShrinking(bool shrinking){
-		m_shrink = shrinking;
-		if(!shrinking)
-			unshrink();
-	}
-
-	/// \brief Define the initial solution for the iterative solver.
-	///
-	/// This method can be used to warm-start the solver. It requires a
-	/// feasible solution (alpha) and the corresponding gradient of the
-	/// dual objective function.
-	void setInitialSolution(RealVector const& alpha, RealVector const& gradient, RealVector const& gradientEdge)
-	{
-		std::size_t n = dimensions();
-		SIZE_CHECK(alpha.size() == n);
-		SIZE_CHECK(gradient.size() == n);
-		for (std::size_t i=0; i<n; i++)
-		{
-			std::size_t j = this->permutation(i);
-			SHARK_ASSERT(alpha(j) >= boxMin(j) && alpha(j) <= boxMax(j));
-			this->m_problem.alpha(i) = alpha(j);
-			this->m_gradient(i) = gradient(j);
-			m_gradientEdge(i) = gradientEdge(j);
-			this->updateAlphaStatus(i);
-		}
-	}
-
-	/// \brief Define the initial solution for the iterative solver.
-	///
-	/// This method can be used to warm-start the solver. It requires a
-	/// feasible solution (alpha), for which it computes the gradient of
-	/// the dual objective function. Note that this is a quadratic time
-	/// operation in the number of non-zero coefficients.
-	void setInitialSolution(RealVector const& alpha)
-	{
-		std::size_t n = dimensions();
-		SIZE_CHECK(alpha.size() == n);
-		RealVector gradient = this->m_problem.linear;
-		RealVector gradientEdge = this->m_problem.linear;
-		blas::vector<QpFloatType> q(n);
-		std::vector<std::size_t> inverse(n);
-		for (std::size_t i=0; i<n; i++) inverse[this->permutation(i)] = i;
-		for (std::size_t i=0; i<n; i++)
-		{
-			double a = alpha(i);
-			if (a == 0.0) continue;
-			this->m_problem.quadratic.row(i, 0, n, q.raw_storage().values);
-			noalias(gradient) -= a * q;
-			std::size_t j = inverse[i];
-			if (a == boxMin(j) || a == boxMax(j)) gradientEdge -= a * q;
-		}
-		setInitialSolution(alpha, gradient, gradientEdge);
-	}
-
-	/// \brief Scales all box constraints by a constant factor and adapts the solution by scaling it by the same factor.
-	void scaleBoxConstraints(double factor, double variableScalingFactor){
-		base_type::scaleBoxConstraints(factor,variableScalingFactor);
-		if(factor != variableScalingFactor){
-			for(std::size_t i = 0; i != dimensions(); ++i){
-				m_gradientEdge(i) = linear(i);
-			}	
-		}
-		else{
-			for(std::size_t i = 0; i != dimensions(); ++i){
-				m_gradientEdge(i) -= linear(i);
-				m_gradientEdge(i) *= factor;
-				m_gradientEdge(i) += linear(i);
-			}
-		}
-	}
-	
-	virtual void deactivateVariable(std::size_t i){
-		SIZE_CHECK(i < dimensions());
-		double alphai = alpha(i);
-		base_type::deactivateVariable(i);
-		updateGradientEdge(i,alphai,0.0);
-	}
-	
-	/// \brief adapts the linear part of the problem and updates the internal data structures accordingly.
-	virtual void setLinear(std::size_t i, double newValue){
-		m_gradientEdge(i) -= linear(i);
-		m_gradientEdge(i) += newValue;
-		base_type::setLinear(i,newValue);
-	}
-	
-	///\brief swap indizes (i,j)
-	void flipCoordinates(std::size_t i,std::size_t j){
-		base_type::flipCoordinates(i,j);
-		std::swap( m_gradientEdge[i], m_gradientEdge[j]);
-	}
-protected:
-	///\brief Updates the edge-part of the gradient when an alpha valu was changed
-	///
-	/// This function overwite the base class method and is called, whenever the
-	/// base class changes an alpha value.
-	void updateGradientEdge(std::size_t i, double oldAlpha, double newAlpha){
-		SIZE_CHECK(i < active());
-		if(!m_shrink || oldAlpha==newAlpha) return;
-		bool isInsideOld = oldAlpha > boxMin(i) && oldAlpha < boxMax(i);
-		bool isInsideNew = newAlpha > boxMin(i) && newAlpha < boxMax(i);
-		//check if variable is relevant at all, that means that old and new alpha value are inside
-		//or old alpha is 0 and new alpha inside
-		if( (oldAlpha == 0 || isInsideOld) && isInsideNew  )
-			return;
-
-		//compute change to the gradient
-		double diff = 0;
-		if(!isInsideOld)//the value was on a border, so remove it's old influeence to the gradient
-			diff -=oldAlpha;
-		if(!isInsideNew){//variable entered boundary or changed from one boundary to another
-			diff  += newAlpha;
-		}
-
-		QpFloatType* q = quadratic().row(i, 0, dimensions());
-		for(std::size_t a = 0; a != dimensions(); ++a){
-			m_gradientEdge(a) -= diff*q[a];
-		}
-	}
-private:
-
-	bool testShrinkVariable(std::size_t a, double largestUp, double smallestDown)const{
 		if (
 			( isLowerBound(a) && gradient(a) < smallestDown)
 			|| ( isUpperBound(a) && gradient(a) >largestUp)
@@ -613,25 +398,13 @@ private:
 		}
 		return false;
 	}
+};
 
-	void getMaxKKTViolations(double& largestUp, double& smallestDown, std::size_t maxIndex){
-		largestUp = -1e100;
-		smallestDown = 1e100;
-		for (std::size_t a = 0; a < maxIndex; a++){
-			if (!isLowerBound(a))
-				smallestDown = std::min(smallestDown,gradient(a));
-			if (!isUpperBound(a))
-				largestUp = std::max(largestUp,gradient(a));
-		}
-	}
-	
-	bool m_isUnshrinked;
-	
-	///\brief true if shrinking is to be used.
-	bool m_shrink;
-
-	///\brief Stores the gradient of the alpha dimeensions which are either 0 or C
-	RealVector m_gradientEdge;
+template<class Problem>
+class BoxConstrainedShrinkingProblem: public BoxBasedShrinkingStrategy<BoxConstrainedProblem<Problem> >{
+public:
+	BoxConstrainedShrinkingProblem(Problem& problem, bool shrink=true)
+	:BoxBasedShrinkingStrategy<BoxConstrainedProblem<Problem> >(problem,shrink){}
 };
 
 }
